Simplified and added confirmation promt if risky alias behavior.
Improved help descriptions and made them more standardized. Parser generation is not in fts.core.parser and rewritten. Allowed --logfile, --verbose, and --quiet flags to be used anywhere in the command. Added a promt to the alias command when a warning is triggered; this may be bypassed with -y or --yes.